Move os modelos de journal relacionados com coleção para a app collection
