+++ journal/models.py
@@ -2,7 +2,6 @@ from django.db import models
 from django.utils.translation import gettext_lazy as _
 
 from core.models import CommonControlField
-from collection.models import Collection
 from .forms import OfficialJournalForm
 
 from wagtail.core.models import Orderable
@@ -30,43 +29,6 @@ class OfficialJournal(CommonControlField):
     ISSNL = models.CharField(_('ISSNL'), max_length=9, null=True, blank=True)
 
     base_form_class = OfficialJournalForm
-
-
-class JournalInCollection(CommonControlField):
-    """
-    Class that represent the Official Journal
-    """
-
-    # TODO futuramente ter um formulário para gerir os dados
-
-    def __unicode__(self):
-        return u'%s' % self.scielo_issn
-
-    def __str__(self):
-        return u'%s' % self.scielo_issn
-
-    scielo_issn = models.CharField(_('SciELO ISSN'), max_length=9, null=False, blank=False)
-    collection = models.ForeignKey(Collection, on_delete=models.CASCADE)
-
-    # TODO acrescentar
-    # data de entrada
-    # data de saída
-    # motivo da saída
-
-
-class SciELOJournal(CommonControlField):
-    """
-    Class that represent the Official Journal
-    """
-
-    def __unicode__(self):
-        return u'%s %s' % (self.official_journal.title, [c.scielo_issn for c in self.collections])
-
-    def __str__(self):
-        return u'%s %s' % (self.official_journal.title, [c.scielo_issn for c in self.collections])
-
-    official_journal = models.ForeignKey(OfficialJournal, on_delete=models.CASCADE)
-    collections = models.ManyToManyField(JournalInCollection)
 
 
 class NonOfficialJournalTitle(ClusterableModel, CommonControlField):
